requirements.txt
torch==2.4.1+cu118
torchvision==0.19.1+cu118

numpy==2.1.2
opencv-python==4.10.0.84
Pillow==10.4.0
PyYAML==6.0.2
tqdm==4.66.5
thop==0.1.1.post2209072238

python=3.10
